pyproject.toml
[tool.poetry]
name = "pyapp-window"
version = "2.0.0a0"
description = "Launch URL with native window in WebView mode. Multiple backend implementations are supported."
authors = ["likianta <likianta@foxmail.com>"]
readme = "README.md"
packages = [{ include = "pyapp_window" }]

[tool.poetry.dependencies]
python = "^3.8"
argsense = "^0.5.8"
lk-utils = "^2.10.3"
psutil = "*"
pywebview = { version = "^5.1", platform = "win32", python = ">=3.8,<3.13" }
# toga = "^0.4.0"
# toga = [
#     { version = "^0.4.0", platform = "darwin", python = ">=3.8,<3.13" },
#     # { version = "^0.4.0", platform = "win32", python = ">=3.8,<3.13" },
# ]
# webui2 = "*"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
